+++ test/download_utils_test.dart
@@ -115,6 +115,18 @@ void main() {
       expect(status.getProblemCode(packageIsKnownInternal: true), isNull);
     });
 
+    test('false positive: internal #1', () async {
+      final status =
+          await UrlChecker().hasExternalHostname(Uri.parse('https://pubxdev/'));
+      expect(status, isTrue);
+    });
+
+    test('false positive: internal #2', () async {
+      final status = await UrlChecker()
+          .hasExternalHostname(Uri.parse('https://onepub.dev/'));
+      expect(status, isTrue);
+    });
+
     test('problem: insecure', () async {
       final status = await UrlChecker().checkStatus('http://pub.dev/');
       expect(status.isInvalid, false);